requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
aiohttp==3.8.1
aiosignal==1.2.0; python_version >= '3.6'
async-timeout==4.0.2; python_version >= '3.6'
attrs==21.4.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
certifi==2022.5.18.1; python_version >= '3.6'
charset-normalizer==2.0.12; python_full_version >= '3.5.0'
frozenlist==1.3.0; python_version >= '3.7'
git+https://github.com/Pycord-Development/pycord@8a4a416b5a41d50fabbd1a2e1832064a56356ee0#egg=py-cord
git+https://github.com/colinhartigan/valclient.py@0dcff9e384943a2889e6b3f8e71781c9fc950bce#egg=valclient
idna==3.3; python_full_version >= '3.5.0'
multidict==6.0.2; python_version >= '3.7'
pillow==9.1.1
pymongo==4.1.1
requests==2.28.0
urllib3==1.26.9; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
yarl==1.7.2; python_version >= '3.6'
